+++ packages/ai/src/generate-text/core-events.ts
