+++ packages/ai/src/generate-text/index.ts
@@ -71,4 +71,4 @@ export type {
   OnStepStartEvent,
   OnToolCallFinishEvent,
   OnToolCallStartEvent,
-} from './callback-events';
+} from './core-events';
